Added process_move.php split from logic into seperate file
