+++ src/chessgame.php
@@ -9,6 +9,8 @@
 </head>
 <body>
 <?php
+    require_once("logic/process_move.php");
+
     require_once("logic/logic.php");
 
     if( isset($_GET['gamemode'])){
