+++ src/logic/logic.php
@@ -9,31 +9,6 @@ require_once("chessboard.php");
 require_once("bot.php");
 
 
-session_start();
-
-if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    $move = get_played_move();
-    $selected_piece_x = $move[0];
-    $selected_piece_y = $move[1];
-    $move_to_x = $move[2];
-    $move_to_y = $move[3];
-    $_SESSION['chess_game']->input_move($selected_piece_x,$selected_piece_y,$move_to_x,$move_to_y);
-}
-
-
-
-function get_played_move() : string 
-{
-    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-        $move = isset($_POST['move_to_id']) ? $_POST['move_to_id'] : '';
-        if(isset($move)){
-            $move=$_POST['selected_piece_id'].$_POST['move_to_id'];
-        }else{
-            $move = 'Error while processing the move';
-        }
-            return $move;
-    }
-}
 
 class Logic
 {
@@ -54,6 +29,15 @@ class Logic
     }
 
 
+    function test($x,$y):void
+    {
+        if($this->chessboard[$x][$y]==""){
+            
+        }else{
+            print("PIECE ON THIS SQUARE:".$x." ".$y."\n");
+        }
+    }
+
     function input_move(int $current_x, int $current_y, int $move_to_x, int $move_to_y):void
     {   
         if($this->chessboard[$current_x][$current_y]==""){
@@ -68,22 +52,16 @@ class Logic
                 $this->whitesturn = !$this->whitesturn; # swap turns
                 $this->is_check($this->chessboard);
                 $this->is_checkmate($this->chessboard);
-                if($this->gamemode=="solo"&&$this->whitesturn==false){
-                    $status = json_decode($this->gamestatus_json, true);
-                    $status['weight'] = evaluate_board($this->chessboard);
-                    $status['botmove'] = minimax($this->chessboard_obj,$this->chessboard, 3, 0, true);
-                    $this->gamestatus_json = json_encode($status);
-                }
                 echo $this->gamestatus_json;
         }
     }
 
 
     #just for testing purpose
-    function debug_output_board(mixed $chessboard): void
+    function debug_output_board(): void
     {
         echo "<pre>";
-        print_r($chessboard);
+        print_r($this->chessboard);
         echo "</pre>";
     }
